windows is hard...
Found several interesting things while testing on windows in vscode.  Unfortunately my environment for testing in vscode on windows consists of packaging a vsix and downloading and installing on windows machine.  Since vsce is strict about checking that dependencies exist and are published to npm, it means to make a change to the underlying code like git-temporal/commons I have to publish commons and publish git-diff-scraper and....  :thinkingface: Need to see if packaging git-temporal-vscode in webpack would allow me to pull in org deps via links link lerna bootstrap creates.

- version bump train / groundhog day
- fix issue with git show working differently on windows
- better logging of exec sync
- better error handling